remove boilerplate and move App.tsx
