+++ src/App.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { SafeAreaView } from 'react-native';
+
+const App = () => {
+  return <SafeAreaView />;
+};
+
+export default App;
